Fixed bug where breakpoints were not reset properly when restarting pdb without closing
